# car_service_back/api-documentation.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/services":{"get":{"tags":["services-controller"],"operationId":"serviceList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Service"}}}}}}},"put":{"tags":["services-controller"],"operationId":"updateService","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Service"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Service"}}}}}}},"/reservations":{"get":{"tags":["reservations-controller"],"operationId":"reservationList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}}}},"put":{"tags":["reservations-controller"],"operationId":"updateReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}}},"/repair-jobs":{"get":{"tags":["repair-job-controller"],"operationId":"repairJobList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RepairJob"}}}}}}},"put":{"tags":["repair-job-controller"],"operationId":"updateRepairJob","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepairJob"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RepairJob"}}}}}}},"/mechanics":{"get":{"tags":["mechanics-controller"],"operationId":"mechanicList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Mechanic"}}}}}}},"put":{"tags":["mechanics-controller"],"operationId":"updateMechanic","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Mechanic"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mechanic"}}}}}}},"/clients":{"get":{"tags":["client-controller"],"operationId":"clientList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Client"}}}}}}},"put":{"tags":["client-controller"],"operationId":"updateClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}}},"/cars":{"get":{"tags":["car-controller"],"operationId":"carList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Car"}}}}}}},"put":{"tags":["car-controller"],"operationId":"updateCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Car"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Car"}}}}}}},"/services/new":{"post":{"tags":["services-controller"],"operationId":"createService","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Service"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Service"}}}}}}},"/reservations/new":{"post":{"tags":["reservations-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}}},"/repair-jobs/new":{"post":{"tags":["repair-job-controller"],"operationId":"createRepairJob","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepairJob"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RepairJob"}}}}}}},"/mechanics/new":{"post":{"tags":["mechanics-controller"],"operationId":"createMechanic","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Mechanic"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mechanic"}}}}}}},"/clients/new":{"post":{"tags":["client-controller"],"operationId":"createClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}}},"/cars/new":{"post":{"tags":["car-controller"],"operationId":"createCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Car"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Car"}}}}}}},"/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/authenticate":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/services/{id}":{"get":{"tags":["services-controller"],"operationId":"getService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Service"}}}}}},"delete":{"tags":["services-controller"],"operationId":"deleteService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/reservations/{id}":{"get":{"tags":["reservations-controller"],"operationId":"getReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}},"delete":{"tags":["reservations-controller"],"operationId":"deleteReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/reservations/schedule":{"get":{"tags":["reservations-controller"],"operationId":"getReservationsSchedule","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}}}}},"/reservations/filter":{"get":{"tags":["reservations-controller"],"operationId":"filterReservations","parameters":[{"name":"filter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ReservationFilter"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}}}}},"/repair-jobs/{id}":{"get":{"tags":["repair-job-controller"],"operationId":"getRepairJob","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RepairJob"}}}}}},"delete":{"tags":["repair-job-controller"],"operationId":"deleteRepairJob","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/mechanics/{id}":{"get":{"tags":["mechanics-controller"],"operationId":"getMechanic","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mechanic"}}}}}},"delete":{"tags":["mechanics-controller"],"operationId":"deleteMechanic","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/clients/{id}":{"get":{"tags":["client-controller"],"operationId":"getClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}},"delete":{"tags":["client-controller"],"operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/cars/{id}":{"get":{"tags":["car-controller"],"operationId":"getCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Car"}}}}}},"delete":{"tags":["car-controller"],"operationId":"deleteCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Service":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"serviceName":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"estimatedDuration":{"type":"integer","format":"int64"}}},"Car":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"model":{"type":"string"},"make":{"type":"string"},"year":{"type":"integer","format":"int32"},"licensePlate":{"type":"string"},"owner":{"$ref":"#/components/schemas/Client"}}},"Client":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"Mechanic":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"specialty":{"type":"string"},"experience":{"type":"integer","format":"int64"}}},"Reservation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"client":{"$ref":"#/components/schemas/Client"},"car":{"$ref":"#/components/schemas/Car"},"mechanic":{"$ref":"#/components/schemas/Mechanic"},"service":{"$ref":"#/components/schemas/Service"},"dateAdded":{"type":"string","format":"date-time"},"visitDateTime":{"type":"string","format":"date-time"},"status":{"type":"string"},"additionalDetails":{"type":"string"}}},"RepairJob":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"client":{"$ref":"#/components/schemas/Client"},"mechanic":{"$ref":"#/components/schemas/Mechanic"},"startDateTime":{"type":"string","format":"date-time"},"endDateTime":{"type":"string","format":"date-time"},"service":{"$ref":"#/components/schemas/Service"},"status":{"type":"string"},"additionalDetails":{"type":"string"}}},"RegistrationRequest":{"required":["name","password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ReservationFilter":{"type":"object","properties":{"dateFrom":{"type":"string","format":"date-time"},"dateTo":{"type":"string","format":"date-time"},"clientId":{"type":"integer","format":"int64"},"carId":{"type":"integer","format":"int64"},"serviceId":{"type":"integer","format":"int64"},"statusId":{"type":"integer","format":"int64"},"mechanicId":{"type":"integer","format":"int64"}}}}}}