# car_service_back/src/main/java/com/example/carrepairshop/model/AuthResponse.java
package com.example.carrepairshop.model;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class AuthResponse {
    private Long userId;
    private String username;
}