# car_service_back/src/main/java/com/example/carrepairshop/model/Mechanic.java
package com.example.carrepairshop.model;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.Data;

@Data
@Entity
@Table(name = "mechanics", schema = "car_service")
public class Mechanic {
    @Id
    @GeneratedValue
    private Long id;
    private String name;
    private String phone;
    private String email;
    private String password;
    private String specialty;
    private Long experience;
}